Comentarios y corrección de errores

+++ src/co/edu/udea/iw/bl/impl/LoanBLImpl.java
@@ -17,8 +17,8 @@ import co.edu.udea.iw.dto.User;
 import co.edu.udea.iw.exception.MyException;
 
 /**
- * Implementacion de la logica de negocio para los prestamos
- * @author Raul Antonio Martinez - rantonio.martinez@udea.edu.co
+ * Implementación de la lógica de negocio para los préstamos
+ * @author Raul Antonio Martinez Silgado - rantonio.martinez@udea.edu.co
  * @author Juan Diego Goez Durango - diego.goez@udea.edu.co
  * @version 2.0
  */
@@ -28,12 +28,32 @@ public class LoanBLImpl implements LoanBL {
 	private LoanDAO loanDAO;
 	private UserDAO userDAO;
 	private DeviceDAO deviceDAO;
-
+	
+	/**
+	 * Obtiene la lista de todos los préstamos que existan en la base de datos
+	 * @return Lista de Préstamos
+	 * @see co.edu.udea.iw.bl.LoanBL#getLoans()
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public List<Loan> getLoans() throws MyException {
 		return loanDAO.getLoans();
 	}
-
+	
+	/**
+	 * Inserta un préstamo en la base de datos con la información ingresada como parámetro
+	 * @param username
+	 * @param startDate
+	 * @param endDate
+	 * @param returnDate
+	 * @param status
+	 * @param code
+	 * @param copy
+	 * @see co.edu.udea.iw.bl.LoanBL#registerLoan(String, Date, Date, Date, String, String, String)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public void registerLoan(String username, Date startDate, Date endDate, Date returnDate, String status, String code,
 			String copy) throws MyException {
@@ -73,6 +93,16 @@ public class LoanBLImpl implements LoanBL {
 		loanDAO.registerLoan(loan);
 	}
 
+	/**
+	 * Obtiene los préstamos para el día(date) ingresado en los que el código, el nombre sea igual a los datos enviados como parámetros
+	 * @param code
+	 * @param copy
+	 * @param date
+	 * @return Lista de Préstamos
+	 * @see co.edu.udea.iw.bl.LoanBL#getLoans(String, String, Date)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public List<Loan> getLoans(String code, String copy, Date date) throws MyException {
 		if (code == null || "".equals(code)) {
@@ -88,7 +118,18 @@ public class LoanBLImpl implements LoanBL {
 		deviceId = new DeviceId(code, copy);		
 		return loanDAO.getLoans(deviceId, date);
 	}
-
+	
+	/**
+	 * Obtiene el préstamo para el usuario indicado en la fecha expresada(startDate) 
+	 * @param username
+	 * @param code
+	 * @param copy
+	 * @param startDate
+	 * @return Un Préstamo
+	 * @see co.edu.udea.iw.bl.LoanBL#getLoan(String, String, String, Date)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public Loan getLoan(String username, String code, String copy, Date startDate) throws MyException {
 		if (username==null || "".equals(username)) {
@@ -112,6 +153,19 @@ public class LoanBLImpl implements LoanBL {
 		return loanDAO.getLoan(loanId);
 	}
 
+	/**
+	 * Actualiza la información del préstamo que corresponde a los datos ingresados como parámetro
+	 * @param username
+	 * @param startDate
+	 * @param endDate
+	 * @param returnDate
+	 * @param status
+	 * @param code
+	 * @param copy
+	 * @see co.edu.udea.iw.bl.LoanBL#updateLoan(String, Date, Date, Date, String, String, String)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public void updateLoan(String username, Date startDate, Date endDate, Date returnDate, String status, String code,
 			String copy) throws MyException {
@@ -146,10 +200,24 @@ public class LoanBLImpl implements LoanBL {
 			throw new MyException("El dispositivo no esta registrado");
 		}
 		LoanId loanId = new LoanId(user, device, startDate);
-		Loan loan = new Loan(loanId, endDate, returnDate, status);
+		Loan loan = loanDAO.getLoan(loanId); 
+		if (loan == null) {
+			throw new MyException("El préstamo no esta registrado");
+		}
+		loan.setReturnDate(returnDate);
+		loan.setStatus(status); //Duda
 		loanDAO.updateLoan(loan);
 	}
 
+	/**
+	 * Obtiene la lista de préstamos para un usuario y estado en específico(por ejemplo los préstamos ACTIVOS) 
+	 * @param username
+	 * @param status
+	 * @return Lista de Préstamos
+	 * @see co.edu.udea.iw.bl.LoanBL#getLoansUser(String, String)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public List<Loan> getLoansUser(String username, String status) throws MyException {
 		if (username==null || "".equals(username)) {
@@ -165,6 +233,15 @@ public class LoanBLImpl implements LoanBL {
 		return loanDAO.getLoans(user, status);
 	}
 
+	/**
+	 * Busca los préstamos de acuerdo al tipo y número de identificación ingresados como parámetros
+	 * @param typeId
+	 * @param numberId
+	 * @return Lista de Préstamos
+	 * @see co.edu.udea.iw.bl.LoanBL#getLoansDevice(String, String)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public List<Loan> getLoansDevice(String typeId, String numberId) throws MyException {
 		if (typeId==null || "".equals(typeId)) {
@@ -176,6 +253,16 @@ public class LoanBLImpl implements LoanBL {
 		return loanDAO.getLoans(typeId, numberId);
 	}
 
+	/**
+	 * Elimina el préstamo que corresponde a los datos ingresados como parámetros
+	 * @param username
+	 * @param code
+	 * @param copy
+	 * @param startDate
+	 * @see co.edu.udea.iw.bl.LoanBL#deleteLoan(String, String, String, Date)
+	 * @throws MyException
+	 * Lanzamos nuestra propia exception para manejarla en una capa superior
+	 */
 	@Override
 	public void deleteLoan(String username, String code, String copy, Date startDate) throws MyException {
 		if (username==null || "".equals(username)) {
@@ -203,8 +290,10 @@ public class LoanBLImpl implements LoanBL {
 			throw new MyException("El dispositivo no esta registrado");
 		}
 		LoanId loanId = new LoanId(user, device, startDate);
-		Loan loan = new Loan();
-		loan.setLoanId(loanId);
+		Loan loan = loanDAO.getLoan(loanId);
+		if(loan == null){ //Duda
+			throw new MyException("El Préstamo no esta registrado");
+		}
 		loanDAO.deleteLoan(loan);
 	}
